Added test for notifications.send_sms
